Restructure parser to do one-token lookahead instead of backtracking
It's easy with this grammar. This makes error reporting much better, since any
ParseException is a real parse error, not just a signal to backtrack.

Also modify the grammar to include the braces in the expression-list rule, so it's
possible to distinguish the normal end of the list from a parse error.