Add logical operators to conditions
This does make loop codegen slightly worse. Before, loops were always
arranged so that no matter what the condition was, the condition check
only entailed one jump (falling through either to the loop body or to
the after-loop code).

For example, a `while (x < 0)` could compile to:

```
  Jump check
body:
  [body]
check:
  CopyFrom x
  JumpN body
  // fall through
after:
```

whereas a `while (x >= 0)` could compile to:

```
check:
  CopyFrom x
  JumpN after
  [body]
  Jump check
after:
```

Now, the condition check is always at the top, which sometimes
necessitates a conditional jump over an unconditional one. E.g. in the
case of `while (x < 0)`, this is the output:

```
check:
  CopyFrom x
  JumpN body
  Jump after
body:
  [body]
  Jump check
after:
```

It's not ideal, but I think this is overall the cleaner approach. This
kind of inefficiency should be cleaned up by a smarter jump optimizer
-- actually analyzing and reordering basic blocks -- rather than a
peephole optimization during initial codegen.

Fixes #1