Add the concept of statements
As I wrote in #4, the fact that everything is an expression is kind of
weird, since all the control flow constructs don't formally have
values.

I decided on adding semicolons to enhance the resemblance to C, and
possibly to avoid some ambiguities. (E.g. I experimented with parsing
negative numbers, and it got messy without some kind of separator.)

Fixes #4